[frontend] typescript components  migration
