+++ src/frontend/web_application/src/components/BlockList/index.spec.tsx
