+++ src/frontend/web_application/src/components/Brand/index.tsx
@@ -1,0 +1,42 @@
+import * as React from 'react';
+import classnames from 'classnames';
+import './style.scss';
+
+interface Props {
+  className?: string;
+  theme?: 'low';
+  responsive?: boolean;
+}
+function Brand({ className, theme, responsive = false }: Props) {
+  const brandClassName = classnames(
+    'm-brand',
+    {
+      'm-brand--low': theme === 'low',
+      'm-brand--responsive': responsive,
+    },
+    className
+  );
+
+  return (
+    <div className={brandClassName}>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 350 250"
+        className="m-brand__logo"
+        aria-label="Logo Caliopen"
+      >
+        <path d="M231.7 2c-43.9 0-81.5 26.4-98 64.1A84.1 84.1 0 0 0 97 57.8a85.4 85.4 0 0 0-85.4 85.4 84.4 84.4 0 0 0 17.7 52l2.3 46 43.2-15.5a84.9 84.9 0 0 0 22.3 3 85.3 85.3 0 0 0 69.4-35.7 106 106 0 0 0 65.2 22.3c10.3 0 20.3-1.5 29.7-4.3l53.6 18.4 3.2-58.4a106 106 0 0 0 20.2-62.5A106.8 106.8 0 0 0 231.7 2zm-.3 24a82.1 82.1 0 0 1 82.1 82.1 83.2 83.2 0 0 1-17 49.8l-.7 40.7-38.2-12.7a83 83 0 0 1-26.4 4.3 82 82 0 0 1-72.2-43 62 62 0 0 1-62 58c-6.8 0-13.3-1.1-19.5-3.1L51 211.9l-1.5-29A61.8 61.8 0 0 1 35.1 143a62.1 62.1 0 0 1 62-62 62 62 0 0 1 52.1 28.2v-1.2A82.1 82.1 0 0 1 231.4 26zm-1 24.5A57.4 57.4 0 0 0 173 108a57.4 57.4 0 0 0 57.5 57.5c6.5 0 12.6-1 18.4-3l26.6 8.8-2.7-24.3a57.2 57.2 0 0 0 15.2-39 57.4 57.4 0 0 0-57.5-57.5zM95.7 104a38.8 38.8 0 0 0-28.9 64.7l-.7 18 16.1-7.6a39.5 39.5 0 0 0 13.6 2.5 38.8 38.8 0 1 0 0-77.6z" />
+      </svg>
+      <svg
+        xmlns="http://www.w3.org/2000/svg"
+        viewBox="0 0 925 250"
+        className="m-brand__brand-name"
+        aria-label="Caliopen"
+      >
+        <path d="M251 100.2a43 43 0 0 0-3.6-7.2 59 59 0 0 0-48.6-22.2h-5.4c-7 0-14.7 1.6-22.9 4.8-22.7 12-34 30-34 54.2 0 5.9 1.2 12.6 3.4 20.3 10.5 25.8 28.9 38.7 55 38.7h3.8a67.6 67.6 0 0 0 31-7.7c1 3.9 4.6 6.6 10.3 7.8h2c6.4 0 10.1-4.3 11-13V107c-.2-3-1-5.1-2-6.9zM198.4 165h-3.8c-4.8 0-9-.8-12.7-2.4-14.3-6.4-21.5-17.2-21.5-32.4l1-9.7c5.5-17.4 16.5-26 33-26h4c13 0 23 4.1 30 12.5l1 5v35a32.7 32.6 0 0 0-1 4.9c-6.3 8.8-16.3 13-30 13zm90-165h2c7.1 1.4 10.7 5 10.7 10.6v135.5c0 10 9.4 16.3 28.2 18.9 9.2 0 14.5 2.6 16 7.7l.6 2.9v2.9c0 5.1-3.6 8.5-10.7 10.1h-4.3a61 61 0 0 1-46.8-20.8 38.8 38.8 0 0 1-6.8-24.7V15.5c0-8 2.4-12.8 7.3-14.5zm83.8 0h2c6.2 0 10 4 11 12-1.5 7.8-5 11.6-10.6 11.6h-2.4c-7.1-1.5-10.7-5-10.7-10.6v-2.4C363 3.6 366.5 0 372.1 0zm0 70.7h2.9c5.2 0 8.5 4 10.2 12v92.9c-1 8.7-4.7 13-11.2 13h-1.9c-7.1-1.5-10.7-5-10.7-10.6V81.3c1.5-7 5.1-10.6 10.7-10.6zm86.3 0h4.4c19 0 35 10 47.7 30a54.5 54.5 0 0 1 6.4 24.6v8.7c0 19.3-9.8 35.3-29.3 48-8.9 4.4-17.5 6.7-25.7 6.7h-3.4c-20.8 0-37.3-10.8-49.6-32.3a62 62 0 0 1-5.4-26.2v-1c0-27.7 13.8-46.6 41.4-56.5a48 48 0 0 1 13.5-2zm-31.1 57v2.9c0 17.7 9 29 26.7 33.9l4.4.5h3.4c13.7 0 23.8-8 30.2-23.8l1-7.2v-8.8c0-13.7-8-23.7-24-30l-6.3-.9h-4.4c-13.2 0-23 7.6-29.7 22.7-.8 4.1-1.3 7.7-1.3 10.7zm163.7-58c28.1.9 46.9 13.3 56.4 37.3a48.9 48.9 0 0 1 3.8 18.3v6.8c0 20.6-10.2 37.2-30.6 49.8a67.2 67 0 0 1-28.3 6.3h-33.5V239c-1.6 7.4-5.1 11.1-10.7 11.1h-2.4c-5.5 0-9-3.7-10.7-11.1V111.7c0-21.3 15.1-35 45.3-41zm-32.1 40.6v54.2h34c16.7 0 27.7-8 33-24.2 1-2.6 1.5-5.4 1.5-8.3v-2.4l.5-.4-.5-.5v-1.4c0-16.6-8.7-27.5-26.2-33l-9.3-1c-15.3 0-26.1 4.7-32.6 14zM723 70.7h5.4c23.7 0 41 9.3 52.1 28a58.6 58.4 0 0 1 2.4 12c0 7.4-5.6 11.7-17 13.1l-74.4 15.5c0 5.4 4.2 12 12.7 19.8a40.8 40.8 0 0 0 20 5.9h3.8a50 50 0 0 0 28.3-9.2l4.9-1c7.5 1.7 11.2 5.4 11.2 11.1v2.4c0 7.5-11 14-32.6 19.4l-11.7 1h-3.3c-25.5 0-43.7-12.5-54.5-37.3a66 66 0 0 1-3.9-21.8c0-24.3 11.4-42.4 34-54.1a61 61 0 0 1 22.6-4.8zm-30.1 44l60.3-12.5v-.6a47.5 47.5 0 0 0-24.3-7.3H724a32.3 32.3 0 0 0-31.1 20.4zm159.6-44c22 0 38.8 8.4 50.6 25.1 2.3 5.5 3.3 9.7 3.3 12.6V178c-1.5 7-5 10.6-10.6 10.6h-2c-6.2 0-10-4.2-11.1-12.5v-66.3c0-3.7-3.7-7.3-11.1-11a43.4 43.4 0 0 0-19.5-4.4h-3.8c-13.7 0-23.7 4.3-30.2 13l-1 3.9V178c-1.6 7-5.1 10.6-10.6 10.6h-2c-6.4 0-10.1-4.3-11-13v-62.4c0-23.2 16.4-37.4 49.5-42.6h9.5zm-771.7 0h-5.3c-7 0-14.7 1.6-23 4.8-22.6 11.8-34 30-34 54.2 0 7 1.4 14.3 3.9 21.8 10.9 24.8 29 37.2 54.5 37.2h3.4l11.6-1c21.7-5.3 32.6-11.8 32.6-19.3V166c0-5.8-3.8-9.5-11.2-11-3 0-4.9.9-4.9.9-9 6.1-18.4 9.2-28.2 9.2h-3.9a40 40 0 0 1-20-5.9c-8.4-7.8-12.7-14.4-12.7-19.8-1.9-16 1.5-24.7 1.5-24.7a32.2 32.2 0 0 1 31.1-20.3h5c8 0 16 2.4 24.2 7.3l1.7 1s1.8 1 4.9 1c7.5-1.6 11.2-5.4 11.2-11.1V90c0-7.5-11.3-13.5-33.1-18.3 0 0-2.1-.8-9.3-1.1z" />
+      </svg>
+    </div>
+  );
+}
+
+export default Brand;
