--- src/frontend/web_application/src/components/CollectionFieldGroup/components/DeleteFieldGroup/style.scss
@@ -1,13 +1,0 @@
-@import '../../../../styles/vendor/bootstrap_foundation-sites';
-
-.m-delete-field-group {
-  @include flex-grid-row;
-
-  &__input {
-    @include flex-grid-size;
-  }
-
-  &__button {
-    @include flex-grid-size(shrink);
-  }
-}
