--- src/frontend/web_application/src/components/CollectionFieldGroup/index.js
@@ -1,3 +1,0 @@
-import Presenter from './presenter';
-
-export default Presenter;
