--- src/frontend/web_application/src/components/CollectionFieldGroup/style.scss
@@ -1,7 +1,0 @@
-@import '../../styles/common';
-
-.m-collection-field-group {
-  &__delete-group {
-    margin-top: map_get($co-form__spacing, 'small');
-  }
-}
